hard start/stop times to video player
This pull request makes the video player module respect
the video start/stop times specified by the instructor
as a hard start/stops, i.e. the video treats the start
time as the beginning of the video and the stop time as
the end.

Previously, the video module displayed the full
video and displayed a jquery ui slider range representing
the start and end times.

This new functionality invalidates two previous acceptance
tests for the video module:

    - slider visible test
    - finish time video acceptance test.

These tests are removed from the common acceptance tests.